requirements.txt
# Flask 核心套件
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-Mail==0.9.0
Flask-Cors==5.0.0
flask-paginate==2023.10.24
Flask-SocketIO==5.3.6

# 資料庫相關
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
alembic==1.15.2

# 工具和輔助套件
python-dotenv==1.0.1
Werkzeug==3.0.1
Jinja2==3.1.2
email-validator==2.1.0.post1
gunicorn==21.2.0
click==8.1.7
blinker==1.7.0
eventlet==0.35.2

# 安全性相關
cryptography==42.0.5
itsdangerous==2.1.2
MarkupSafe==2.1.3
python-jose==3.3.0
